requirements.txt
djangorestframework~=3.12.2
Django~=3.1.7
pandas~=1.2.2
requests~=2.25.1
selenium~=3.141.0
lighthouse~=1.0.2
numpy~=1.19.5
sklearn~=0.0
scikit-learn~=0.24.1
matplotlib~=3.3.4
Pillow~=8.1.0